+++ pygsti/modelmembers/povms/computationaleffect.py
@@ -83,11 +83,6 @@ class ComputationalBasisPOVMEffect(_POVMEffect):
         -------
         ComputationalBasisPOVMEffect
         """
-        #if evotype in ('stabilizer', 'statevec'):
-        #    nqubits = int(round(_np.log2(len(vec))))
-        #    v0 = _np.array((1, 0), complex)  # '0' qubit state as complex state vec
-        #    v1 = _np.array((0, 1), complex)  # '1' qubit state as complex state vec
-        #else:
         nqubits = int(round(_np.log2(len(vec)) / 2))
         v0 = 1.0 / _np.sqrt(2) * _np.array((1, 0, 0, 1), 'd')  # '0' qubit state as Pauli dmvec
         v1 = 1.0 / _np.sqrt(2) * _np.array((1, 0, 0, -1), 'd')  # '1' qubit state as Pauli dmvec
